UI fixes changes and add clear all button

+++ app/src/main/java/xyz/nspain/grocerybudget/MainActivity.java
@@ -7,16 +7,17 @@ import android.content.DialogInterface;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.AppCompatImageButton;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
+import android.support.v7.widget.helper.ItemTouchHelper;
 import android.telephony.TelephonyManager;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -25,6 +26,7 @@ import android.view.MenuItem;
 import android.view.SubMenu;
 import android.view.View;
 import android.widget.EditText;
+import android.widget.TextView;
 
 import java.math.BigDecimal;
 import java.text.NumberFormat;
@@ -33,18 +35,22 @@ import java.util.List;
 import xyz.nspain.grocerybudget.persistance.Item;
 import xyz.nspain.grocerybudget.persistance.ShoppingList;
 
+import static xyz.nspain.grocerybudget.AdapterNotifyMessage.*;
+
 public class MainActivity extends AppCompatActivity {
 
     private RecyclerView mRecyclerView;
     private ShoppingListAdapter mShoppingListAdapter;
     private RecyclerView.LayoutManager mLayoutManager;
-    private EditText mTotalCostView;
+    private TextView mTotalCostView;
     private ShoppingListViewModel mShoppingListViewModel;
-    private FloatingActionButton mAddItemFAB;
+    private AppCompatImageButton mAddButton;
+    private AppCompatImageButton mClearAllButton;
     private DrawerLayout mNavigationDrawer;
     private NavigationView mNavigationView;
     private NumberFormat mCurrencyFormatter;
     private ShoppingListEditAdapter mShoppingListEditAdapter;
+    private ItemTouchHelper mItemTouchHelper;
 
     /**
      * Tag for debugging purposes
@@ -54,29 +60,69 @@ public class MainActivity extends AppCompatActivity {
 
     private ShoppingList shoppingList;
     private Toolbar mToolbar;
+    private boolean mIsUpdatedFromView = false;
+
+    private View.OnClickListener mCreateItemOnClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            Item item = new Item("", new BigDecimal(0), false);
+            mShoppingListViewModel.insertItem(item);
+        }
+    };
+
+    private View.OnClickListener mCreateListOnClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            ShoppingList list = new ShoppingList("");
+            mShoppingListViewModel.insertList(list);
+        }
+    };
+
+    private View.OnClickListener mClearListsClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            mShoppingListViewModel.deleteAllLists();
+        }
+    };
+
+    private View.OnClickListener mClearItemsClickListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            mShoppingListViewModel.deleteItemsInCurrentList();
+        }
+    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+        mCurrencyFormatter = NumberFormat.getCurrencyInstance(getCurrentLocale());
+
         mShoppingListAdapter = new ShoppingListAdapter(this, mCurrencyFormatter);
         mShoppingListEditAdapter = new ShoppingListEditAdapter(this);
 
-        mCurrencyFormatter = NumberFormat.getCurrencyInstance(getCurrentLocale());
-        mTotalCostView = findViewById(R.id.totalCost);
+        mTotalCostView = findViewById(R.id.total_cost);
         mShoppingListViewModel = ViewModelProviders.of(this).get(ShoppingListViewModel.class);
         mRecyclerView = findViewById(R.id.shoppingList);
         mLayoutManager = new LinearLayoutManager(this);
-        mAddItemFAB = findViewById(R.id.fab);
+        mAddButton = findViewById(R.id.add_button);
+        mClearAllButton = findViewById(R.id.clear_all_button);
         mNavigationDrawer = findViewById(R.id.main_activity);
         mNavigationView = findViewById(R.id.nav_view);
         mToolbar = findViewById(R.id.toolbar);
 
+        mAddButton.setOnClickListener(mCreateItemOnClickListener);
+        mClearAllButton.setOnClickListener(mClearItemsClickListener);
+
+        ItemTouchHelper.Callback itemTouchHelperCallback = new SimpleItemTouchHelperCallback(mShoppingListEditAdapter);
+        mItemTouchHelper = new ItemTouchHelper(itemTouchHelperCallback);
+        mItemTouchHelper.attachToRecyclerView(mRecyclerView);
+
+
         setupToolbar();
         setupRecyclerView();
         setupViewModelObservers();
-        setupFloatingActionButton();
         setupNavigationDrawer();
     }
 
@@ -126,14 +172,19 @@ public class MainActivity extends AppCompatActivity {
                 menuItem.setChecked(true);
                 mNavigationDrawer.closeDrawers();
 
-                Log.d(TAG, "Selected " + menuItem);
                 if (menuItem.getItemId() == R.id.new_list) {
                     createNewList();
                 } else if (menuItem.getItemId() == R.id.edit_lists) {
+                    mAddButton.setOnClickListener(mCreateListOnClickListener);
+                    mClearAllButton.setOnClickListener(mClearListsClickListener);
                     mRecyclerView.setAdapter(mShoppingListEditAdapter);
+                    mToolbar.setTitle(R.string.list_edit_title);
+                    mTotalCostView.setText("");
                 } else {
                     // Must've clicked a list name
                     mShoppingListViewModel.updateCurrentList(menuItem.getTitle().toString());
+                    mAddButton.setOnClickListener(mCreateItemOnClickListener);
+                    mClearAllButton.setOnClickListener(mClearItemsClickListener);
                     mRecyclerView.setAdapter(mShoppingListAdapter);
                 }
 
@@ -162,7 +213,6 @@ public class MainActivity extends AppCompatActivity {
         alertBuilder.setNegativeButton(R.string.new_list_dialog_neg, new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
-
             }
         });
         alertBuilder.show();
@@ -178,23 +228,23 @@ public class MainActivity extends AppCompatActivity {
             public void onItemRangeChanged(int positionStart, int itemCount, Object payload) {
                 super.onItemRangeChanged(positionStart, itemCount, payload);
                 AdapterNotifyMessage msg = (AdapterNotifyMessage) payload;
-                if (msg != null) {
-                    switch (msg.getChangeType()) {
-                        case DELETE_ITEM:
-                            Log.d(TAG, "Deletion message found, deleting " + msg.getPayload());
-                            mShoppingListViewModel.deleteItem((Item)msg.getPayload());
-                            break;
-                        case INSERT_ITEM:
-                            break;
-                        case UPDATE_ITEM:
-                            Log.d(TAG, "Update message found, updating " + msg.getPayload());
-                            mShoppingListViewModel.updateItem((Item)msg.getPayload());
-                            break;
-                    }
-                } else {
-                    Log.d(TAG, "Payload is null so not sending update to view model");
+                if (msg != null && msg.getChangeType() == ChangeType.UPDATE_ITEM) {
+                    mIsUpdatedFromView = true;
+                    mShoppingListViewModel.updateItem((Item)msg.getPayload());
                 }
-                mShoppingListViewModel.setIsCircularUpdate(true);
+            }
+
+            @Override
+            public void onItemRangeRemoved(int positionStart, int itemCount) {
+                super.onItemRangeRemoved(positionStart, itemCount);
+                RecyclerViewAdapterWithRemovePayload adapter = (RecyclerViewAdapterWithRemovePayload) mRecyclerView.getAdapter();
+                if (adapter != null) {
+                    AdapterNotifyMessage msg = (AdapterNotifyMessage) adapter.getRemovedItemPayload();
+                    if (msg != null && msg.getChangeType() == ChangeType.DELETE_ITEM) {
+                        mIsUpdatedFromView = true;
+                        mShoppingListViewModel.deleteItem((Item)msg.getPayload());
+                    }
+                }
             }
         });
 
@@ -205,14 +255,29 @@ public class MainActivity extends AppCompatActivity {
                 AdapterNotifyMessage msg = (AdapterNotifyMessage) payload;
                 if (msg != null) {
                     switch (msg.getChangeType()) {
-                        case DELETE_LIST:
-                            mShoppingListViewModel.deleteList((ShoppingList) msg.getPayload());
-                            break;
-                        case INSERT_LIST:
-                            break;
                         case UPDATE_LIST:
+                            mIsUpdatedFromView = true;
                             mShoppingListViewModel.updateList((ShoppingList) msg.getPayload());
                             break;
+                        case VIEW_LIST:
+                            mShoppingListViewModel.updateCurrentList(((ShoppingList) msg.getPayload()).getName());
+                            mAddButton.setOnClickListener(mCreateItemOnClickListener);
+                            mRecyclerView.setAdapter(mShoppingListAdapter);
+                            break;
+                    }
+
+                }
+            }
+
+            @Override
+            public void onItemRangeRemoved(int positionStart, int itemCount) {
+                super.onItemRangeRemoved(positionStart, itemCount);
+                RecyclerViewAdapterWithRemovePayload adapter = (RecyclerViewAdapterWithRemovePayload) mRecyclerView.getAdapter();
+                if (adapter != null) {
+                    AdapterNotifyMessage msg = (AdapterNotifyMessage) adapter.getRemovedItemPayload();
+                    if (msg != null && msg.getChangeType() == ChangeType.DELETE_LIST) {
+                        mIsUpdatedFromView = true;
+                        mShoppingListViewModel.deleteList((ShoppingList) msg.getPayload());
                     }
                 }
             }
@@ -223,14 +288,10 @@ public class MainActivity extends AppCompatActivity {
         mShoppingListViewModel.getItems().observe(this, new Observer<List<Item>>() {
             @Override
             public void onChanged(@Nullable List<Item> items) {
-//                if (!mShoppingListViewModel.isCircularUpdate()) {
-                    Log.d(TAG, "Is not circular update");
+                if (!mIsUpdatedFromView) {
                     mShoppingListAdapter.setItems(items);
-                    mShoppingListViewModel.setIsCircularUpdate(true);
-//                } else {
-                    mShoppingListViewModel.setIsCircularUpdate(false);
-                    Log.d(TAG, "Is circular update");
-//                }
+                }
+                mIsUpdatedFromView = false;
             }
         });
 
@@ -252,20 +313,12 @@ public class MainActivity extends AppCompatActivity {
         mShoppingListViewModel.getLists().observe(this, new Observer<List<ShoppingList>>() {
             @Override
             public void onChanged(@Nullable List<ShoppingList> shoppingLists) {
-                mShoppingListEditAdapter.setLists(shoppingLists);
-            }
-        });
-    }
+                if (!mIsUpdatedFromView && shoppingLists != null) {
+                    mShoppingListEditAdapter.setLists(shoppingLists);
+                    mLayoutManager.scrollToPosition(shoppingLists.size() - 1);
 
-    private void setupFloatingActionButton() {
-        // Setup the floating button to add new items to the shopping list
-        mAddItemFAB.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Item item = new Item("", new BigDecimal(0), false);
-                Log.d(TAG, "Inserting " + item + " into db");
-                mShoppingListViewModel.setIsCircularUpdate(false);
-                mShoppingListViewModel.insertItem(item);
+                }
+                mIsUpdatedFromView = false;
             }
         });
     }
